webapp/scanner/serializers.py
from rest_framework import serializers, viewsets

from .models import Member


# ---------- Members
class MemberSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Member
        fields = ('first_name', 'last_name')


class MemberViewSet(viewsets.ModelViewSet):
    queryset = Member.objects.all()
    serializer_class = MemberSerializer
