webapp/requirements.txt
# Django
Django
django-extensions
djangorestframework

# Other
requests
